app/requirements.txt
transformers>=4.45.0
torch>=2.0.0
peft>=0.7.0
gradio>=4.0.0
Pillow>=9.0.0
accelerate>=0.25.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.0
